pyproject.toml
[project]
name = "nonebot-plugin-anipusher"
version = "0.2.0"
description = "NoneBot plugin for pushing ani-rss/emby webhooks to QQ"
authors = [{ name = "AriadusTT", email = "980868243@qq.com" }]
license = { text = "MIT" }
readme = "README.md"
urls = { Repository = "https://github.com/AriadusTT/nonebot-plugin-AniPusher" }
requires-python = ">=3.9, <4.0"
dependencies = [
    "aiohttp>=3.11.14",
    "aiosqlite>=0.21.0",
    "fastapi>=0.115.0",
    "nonebot2>=2.4.0",
    "pydantic>=1.10.21",
    "nonebot_plugin_localstore>=0.7.4"
]
classifiers = [
    "Programming Language :: Python :: 3.9",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = []
plugin_dirs = [""]


[tool.ruff]
line-length = 120
select = ["F", "E", "W", "I"]  # 仅保留基础检查
ignore = ["E402"]

[build-system]
requires = ["setuptools>=65.0.0", "wheel"]  # 替代 uv_build
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.setuptools]
packages = {find = {where = ["."]}}  # 根目录扫描
package-dir = {"" = "."}
package-data = {"*" = ["res/*.png", "res/*.jpg"]}
include-package-data = true